test: add comprehensive test suite for vault and app functionalities
