+++ src-tauri/src/commands/folder_commands.rs
@@ -1,5 +1,4 @@
 use anyhow::Result;
-use std::path::PathBuf;
 
 use crate::errors::AppError;
 use crate::vault::VaultManager;
