+++ src-tauri/src/tests/mod.rs
@@ -1,0 +1,8 @@
+// Test modules
+pub mod app_tests;
+pub mod common;
+pub mod error_tests;
+pub mod folder_tests;
+pub mod integration_tests;
+pub mod note_tests;
+pub mod vault_tests;
